+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- analyze native `cdylib` targets with `--cdylib`
+
 ## [0.0.2](https://github.com/Boshen/cargo-bsize/compare/v0.0.1...v0.0.2) - 2026-08-19
 
 ### Added